requirements.txt
allure_python_commons==2.13.5
pytest==8.3.4
selenium==4.30.0
pytest-xdist==3.6.1